+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/MiniApp.kt
@@ -105,7 +105,7 @@ abstract class MiniApp internal constructor() {
     abstract suspend fun fetchInfo(appId: String): MiniAppInfo
 
     /**
-     * Get custom permissions with grant results per MiniApp from this SDK.
+     * Get only the custom permissions listed in the Mini App's manifest with grant results per MiniApp from this SDK.
      * @param miniAppId mini app id as the key to retrieve data from cache.
      * @return [MiniAppCustomPermission] an object contains the grant results per mini app.
      */
@@ -114,7 +114,7 @@ abstract class MiniApp internal constructor() {
     ): MiniAppCustomPermission
 
     /**
-     * Store custom permissions with grant results per MiniApp inside this SDK.
+     * Store only the custom permissions listed in the Mini App's manifest with grant results per MiniApp inside this SDK.
      * @param miniAppCustomPermission the supplied custom permissions to be stored in cache.
      */
     abstract fun setCustomPermissions(