+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/storage/MiniAppManifestCache.kt
@@ -3,7 +3,6 @@ package com.rakuten.tech.mobile.miniapp.storage
 import android.annotation.SuppressLint
 import android.content.Context
 import android.content.SharedPreferences
-import android.util.Log
 import androidx.annotation.VisibleForTesting
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
@@ -24,12 +23,11 @@ internal class MiniAppManifestCache(context: Context) {
 
     /**
      * Reads the grant results from SharedPreferences.
-     * @param [miniAppId] the key provided to find the stored manifest per MiniApp
-     * @return [MiniAppManifest] an object to contain the manifest per MiniApp
+     * @param [miniAppId] the key provided to find the stored manifest per MiniApp.
+     * @return [MiniAppManifest] an object to contain the manifest per MiniApp.
      * if data has been stored in cache, otherwise empty value.
      */
     @SuppressLint("RestrictedApi")
-    @SuppressWarnings("NestedBlockDepth")
     fun readMiniAppManifest(miniAppId: String): MiniAppManifest {
         val empty = MiniAppManifest(emptyList(), emptyList(), emptyMap())
         return if (doesDataExist(miniAppId)) {
@@ -38,16 +36,12 @@ internal class MiniAppManifestCache(context: Context) {
                     prefs.getString(miniAppId, ""),
                     object : TypeToken<MiniAppManifest>() {}.type
                 )
-                Log.d("AAAAA21111~", ""+miniAppId)
-                Log.d("AAAAA21111", ""+cachedManifest.toString())
                 cachedManifest
             } catch (e: Exception) {
                 // if there is any exception, return the empty value
-                Log.d("AAAAA22", ""+empty.toString())
                 empty
             }
         } else {
-            Log.d("AAAAA23", ""+empty.toString())
             // if value hasn't been found in SharedPreferences, save the empty value
             storeMiniAppManifest(miniAppId, empty)
             empty
@@ -56,15 +50,14 @@ internal class MiniAppManifestCache(context: Context) {
 
     /**
      * Stores [MiniAppManifest] to SharedPreferences.
+     * @param [miniAppId] the key provided to find the stored manifest per MiniApp.
      * @param [miniAppManifest] an object to contain the manifest values per MiniApp.
      */
     fun storeMiniAppManifest(
         miniAppId: String,
         miniAppManifest: MiniAppManifest
     ) {
-        Log.d("AAAAA21111~~", ""+miniAppId)
         val jsonToStore: String = Gson().toJson(miniAppManifest)
         prefs.edit().putString(miniAppId, jsonToStore).apply()
-        Log.d("AAAAAMP cache set",""+jsonToStore)
     }
 }